finished implementing JWTAuth in Meeting, Auth, and Registration

+++ app/Http/Controllers/MeetingController.php
@@ -9,11 +9,15 @@ use App\Http\Requests;
 
 use App\Meeting;
 
+use JWTAuth;
+
 class MeetingController extends Controller
 {
     public function __construct()
     {
-        // $this->middleware('name');
+      $this->middleware('jwt.auth', ['only' => [
+        'store', 'update', 'destroy'
+      ]]);
     }
 
     /**
@@ -49,14 +53,17 @@ class MeetingController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'description' => 'required',
-            'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
+            'time' => 'required|date_format:YmdHie'
         ]);
+
+        if(!$user = JWTAuth::parseToken()->authenticate()){
+          return response()->json(['msg' => 'User not found'], 404);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = new Meeting([
             'time' => Carbon::createFromFormat('YmdHie', $time),
@@ -116,14 +123,17 @@ class MeetingController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'description' => 'required',
-            'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
+            'time' => 'required|date_format:YmdHie'
         ]);
+
+        if(!$user = JWTAuth::parseToken()->authenticate()){
+          return response()->json(['msg' => 'User not found'], 404);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = Meeting::with('users')->findOrFail($id);
 
@@ -159,6 +169,15 @@ class MeetingController extends Controller
     public function destroy($id)
     {
         $meeting = Meeting::findOrFail($id);
+
+        if(!$user = JWTAuth::parseToken()->authenticate()){
+          return response()->json(['msg' => 'User not found'], 404);
+        }
+
+        if (!$meeting->users()->where('users.id', $user_id)->first()) {
+            return response()->json(['msg' => 'user not registered for meeting, update not successful'], 401);
+        };
+
         $users = $meeting->users;
         $meeting->users()->detach();
         if (!$meeting->delete()) {
